Improve, localize, and display expression errors
Errors have been refactored into codes, that can be translated and
interpolated using the new ErrorMessages class in SeExpr2Editor.

As an addition, I have re-enabled number highlighting (it seems there
was a typo in Janet Berlin's work from ages ago), and added a wavy
underline over every error that is added to the editor.

While at it, disable some bits from the Flex parser that are not used in
SeExpr2.