Fix Vec1d promotion with LLVM evaluator
Before this commit, the LLVM evaluator would fail with a prep error if
the script result type was FP(1) and the underlying LLVM type was
VectorType (which are nominally equivalent, but under the hood, one is a
double and the other a pointer to a store of 1 x double).

This commit checks this circumstance and promotes FP(1) to FP(3)
correctly, bringing both interpreter and LLVM evaluator to feature
parity again.